requirements.txt
Flask
tensorflow
Pillow
gunicorn